Disable startup SpringBoot logo. Relocate telegram bot config classes.
