+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
   application:
     name: AnalyticsSpringApplication
-# Spring settings
+  # Spring settings
   datasource:
     url: jdbc:mysql://localhost:3306/smartstorage
     driverClassName: com.mysql.cj.jdbc.Driver
@@ -12,6 +12,8 @@ spring:
       ddl-auto: update
     show-sql: true
     open-in-view: false
+  main:
+    banner-mode: "off"
 
 springdoc:
   api-docs:
@@ -28,7 +30,7 @@ server:
 crypto:
   api:
     key: 3c6ca286-a0b1-4da4-8657-29ee33f2e27c
-# API endpoint url
+  # API endpoint url
   base:
     url: https://pro-api.coinmarketcap.com/v1/cryptocurrency/quotes/latest
